refactor: move determinant object to separate file under numerics
